Sanity check for Tester.saveTestConfigData
Currently the method assumes keys exist in the test config.php which
is not the case when you have setup the environment following the documentation.

https://docs.espocrm.com/development/tests/

For example, the key `lastModifiedTime` is not there, and the method is called
to set it during the setup phase.

This commit simply adds a sanity check of `array_key_exists`.