apply MethodArgumentSpaceFixer
In method arguments and method call, there
MUST NOT be a space before each comma and
there MUST be one space after each comma.
Argument lists MAY be split across multiple lines,
where each subsequent line is indented once.
When doing so, the first item in the list MUST be
on the next line, and there MUST be
only one argument per line