Update Web.Config
JSON wouldn't work on IIS if the MIME entries listed in "web.config" already existed in the "MIME Types" for the website, this update removes and re-adds the "MIME types" which resolves the issue.

Alternatively, manually delete the 3 MIME types against the website and then the original web.config file can be used.